requirements.txt
Django==1.9.9
django-tinymce==2.7.0
pytz==2018.4
